requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.8.2
certifi==2024.8.30
class-registry==2.1.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
filters==1.3.2
Flask==3.0.3
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
looseversion==1.3.0
MarkupSafe==3.0.2
nest-asyncio==1.6.0
pathlib==1.0.1
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.12
python-dateutil==2.9.0.post0
python-telegram-bot==21.6
pytz==2024.2
regex==2024.9.11
six==1.16.0
sniffio==1.3.1
traits==6.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
